requirements.txt
anyio==3.6.2
asttokens==2.0.5
beautifulsoup4==4.12.2
blinker==1.7.0
blis==0.7.9
bs4==0.0.1
catalogue==2.0.8
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.0.4
contourpy==1.2.1
cryptography==41.0.5
cycler==0.11.0
cymem==2.0.7
deprecation==2.1.0
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.5.0/en_core_web_md-3.5.0-py3-none-any.whl#sha256=b492fd2dfeee931dd60232fd1a122bb5c3542ebcd5d614f11844647050aa539e
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl#sha256=0964370218b7e1672a30ac50d72cdc6b16f7c867496f1d60925691188f4d2510
exceptiongroup==1.1.1
executing==0.8.3
fastapi==0.95.2
filelock==3.12.2
Flask==2.3.3
Flask-Cors==4.0.0
fonttools==4.34.4
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.42
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
graphviz==0.20.1
html2text==2020.1.16
huggingface-hub==0.21.3
icecream==2.1.3
idna==3.4
iniconfig==2.0.0
iso-639==0.4.5
iso3166==2.1.1
itsdangerous==2.1.2
jellyfish==0.11.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
langcodes==3.3.0
langdetect==1.0.7
language_data==1.2.0
lxml==4.9.2
marisa-trie==1.2.0
markdownify==0.11.6
MarkupSafe==2.1.2
matplotlib==3.5.3
murmurhash==1.0.9
networkx==3.1
nltk==3.8.1
numpy==1.24.3
packaging==23.1
pandas==2.0.1
pdfminer.six==20221105
pickleshare==0.7.5
pikepdf==8.5.1
pillow==10.3.0
pip==24.0
pluggy==1.0.0
preshed==3.0.8
pycountry-convert==0.7.2
pycparser==2.21
pydantic==1.10.8
Pygments==2.16.1
PyMuPDF==1.23.5
PyMuPDFb==1.23.5
pyparsing==3.0.9
pytest==7.3.1
pytextrank==3.3.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.31.0
safetensors==0.4.2
scikit-learn==1.2.2
scikit-multilearn==0.2.0
scipy==1.10.1
segtok==1.5.11
setuptools==70.0.0
six==1.16.0
smart-open==6.3.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.4.1
spacy==3.5.3
spacy-langdetect==0.1.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
starlette==0.27.0
tabulate==0.9.0
thinc==8.1.10
threadpoolctl==3.1.0
tika==2.6.0
tokenizers==0.15.2
tomli==2.0.1
torch==1.12.0
tqdm==4.65.0
transformers==4.38.1
typer==0.7.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
wasabi==1.1.1
weasel==0.3.4
Werkzeug==2.3.7
wheel==0.43.0
yake==0.4.8
psutil==5.9.1
